+++ app/charts/shared/legend-color.tsx
@@ -40,10 +40,11 @@ import { useChartInteractiveFilters } from "@/stores/interactive-filters";
 import { interlace } from "@/utils/interlace";
 import { makeDimensionValueSorters } from "@/utils/sorting-values";
 import useEvent from "@/utils/use-event";
+import { Icon } from "@/icons";
 
 import { DimensionsById } from "./ChartProps";
 
-export type LegendSymbol = "square" | "line" | "circle";
+export type LegendSymbol = "square" | "line" | "circle" | "cross";
 
 type LegendItemUsage = "legend" | "tooltip" | "colorPicker";
 
@@ -99,7 +100,7 @@ const useItemStyles = makeStyles<Theme, ItemStyleProps>((theme) => {
       "&::before": {
         content: "''",
         position: "relative",
-        display: "block",
+        display: ({ symbol }) => (symbol === "cross" ? "none" : "block"),
         width: `calc(0.5rem * var(--size-adjust, 1))`,
         height: ({ symbol }) =>
           `calc(${["square", "circle"].includes(symbol) ? "0.5rem" : "2px"} * var(--size-adjust, 1))`,
@@ -221,6 +222,7 @@ export const LegendColor = memo(function LegendColor({
               ? [measureLimit.value]
               : [measureLimit.from, measureLimit.to],
           color: configLimit.color,
+          symbol: !configLimit.symbolType ? "line" : configLimit.symbolType,
         }))}
         getColor={colors}
         getLabel={getColorLabel}
@@ -312,7 +314,12 @@ const LegendColorContent = ({
   numberOfOptions,
 }: {
   groups: ReturnType<typeof useLegendGroups>;
-  limits?: { label: string; values: number[]; color: string }[];
+  limits?: {
+    label: string;
+    values: number[];
+    color: string;
+    symbol: LegendSymbol;
+  }[];
   getColor: (d: string) => string;
   getLabel: (d: string) => string;
   getItemDimension?: (dimensionLabel: string) => Measure | undefined;
@@ -391,16 +398,17 @@ const LegendColorContent = ({
                       onToggle={handleToggle}
                       checked={interactive && active}
                       disabled={soleItemChecked && active}
+                      usage="legend"
                     />
                   );
                 })}
                 {isLastGroup && limits
-                  ? limits.map(({ label, values, color }, i) => (
+                  ? limits.map(({ label, values, color, symbol }, i) => (
                       <LegendItem
                         key={i}
                         item={`${label}: ${values.map(formatNumber).join("-")}`}
                         color={color}
-                        symbol="line"
+                        symbol={symbol}
                       />
                     ))
                   : null}
@@ -434,10 +442,12 @@ export const LegendItem = (props: LegendItemProps) => {
     onToggle,
     checked,
     disabled,
-    usage = "legend",
+    usage: _usage,
   } = props;
+  const usage = _usage ?? "legend";
   const classes = useItemStyles({ symbol, color, usage });
-  const shouldBeBigger = symbol === "circle" || usage === "colorPicker";
+  const shouldBeBigger =
+    (symbol === "circle" && _usage !== "legend") || usage === "colorPicker";
 
   return interactive && onToggle ? (
     <MaybeTooltip
@@ -472,7 +482,14 @@ export const LegendItem = (props: LegendItemProps) => {
     <Flex
       data-testid="legendItem"
       className={clsx(classes.legendItem, shouldBeBigger && classes.bigger)}
+      sx={{
+        alignItems: symbol === "cross" ? "center !important" : "flex-start",
+      }}
     >
+      {/* TODO: Use icons instead of ::before when migrating to new CI / CD */}
+      {symbol === "cross" ? (
+        <Icon size={16} name="close" color={color} />
+      ) : null}
       {dimension ? (
         <OpenMetadataPanelWrapper component={dimension}>
           {/* Account for the added space, to align the symbol and label. */}