+++ CHANGELOG.md
@@ -17,6 +17,7 @@ You can also check the
     go beyond the data range
   - Limits are now also rendered when the axis dimension is a single filter
   - It's now possible to display limits in map charts
+  - Added a way to set different limit symbols for area and line charts
   - Interactive filters can now be set to be opened by default
   - Statistics page is now linked in the footer
 - Fixes